# specs/master/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 网络小说智能问答系统 API
  description: |
    基于RAG架构的网络小说智能问答系统后端API。
    支持小说管理、智能问答、知识图谱可视化、Token统计等功能。
  version: 1.0.0
  contact:
    name: Development Team
    email: dev@novel-qa.com
servers:
  - url: http://localhost:8000/api
    description: 本地开发环境（推荐）
  - url: http://localhost:8000
    description: 本地开发环境（不带前缀，已废弃）
...
paths:
...
  /query:
    post:
      tags: [query]
      summary: 提交查询（非流式）
      description: 提交问题并等待完整结果返回。推荐使用WebSocket流式接口获得更好体验。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: 成功返回查询结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: 小说不存在或未索引完成
        '500':
          $ref: '#/components/responses/InternalError'
  /query/history:
    get:
      tags: [query]
      summary: 获取查询历史
      parameters:
        - name: novelId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 成功返回查询历史
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/QueryHistoryItem'
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
  /query/{queryId}:
    get:
      tags: [query]
      summary: 获取单次查询详情
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功返回查询详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /query/{queryId}/feedback:
    post:
      tags: [query]
      summary: 提交用户反馈
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [feedback]
              properties:
                feedback:
                  type: string
                  enum: [positive, negative]
                note:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: 反馈提交成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string

  # ==================== 可视化 ====================
...
  /stats/tokens:
    get:
      tags: [stats]
      summary: 获取Token统计
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month, year]
            default: month
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 成功返回Token统计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenStatsReport'
  /stats/tokens/trend:
    get:
      tags: [stats]
      summary: 获取Token使用趋势
      description: 按时间段分组统计Token使用量，用于绘制趋势图
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
          description: 时间段类型
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 365
          description: 返回数据点数量
      responses:
        '200':
          description: 趋势数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  period:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                        total_tokens:
                          type: integer
                        query_count:
                          type: integer
...
components:
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
  schemas:
...
    QueryRequest:
      type: object
      required: [novelId, query]
      properties:
        novelId:
          type: integer
        query:
          type: string
          minLength: 1
          maxLength: 1000
        model:
          type: string
          enum: [GLM-4.5-Flash, GLM-4.5, GLM-4-Plus, GLM-4-Long, GLM-4.6]
          default: GLM-4.5
          description: 智谱AI模型名称（官方名称）
    Citation:
      type: object
      properties:
        chapter_num:
          type: integer
          description: 章节号
        chapter_title:
          type: string
          nullable: true
          description: 章节标题
        text:
          type: string
          description: 引用文本
        score:
          type: number
          nullable: true
          description: 相关性分数
    Contradiction:
      type: object
      properties:
        type:
          type: string
          description: 矛盾类型
        early_description:
          type: string
          description: 早期描述
        early_chapter:
          type: integer
          description: 早期章节
        late_description:
          type: string
          description: 后期描述
        late_chapter:
          type: integer
          description: 后期章节
        analysis:
          type: string
          description: 矛盾分析
        confidence:
          type: string
          enum: [high, medium, low]
          description: 置信度
    TokenStats:
      type: object
      properties:
        total_tokens:
          type: integer
          description: 总Token数
        embedding_tokens:
          type: integer
          default: 0
          description: Embedding消耗的Token
        prompt_tokens:
          type: integer
          default: 0
          description: 提示词Token数
        completion_tokens:
          type: integer
          default: 0
          description: 生成内容Token数
        self_rag_tokens:
          type: integer
          default: 0
          description: Self-RAG验证额外消耗的Token
        by_model:
          type: object
          description: 按模型分类的Token统计
          additionalProperties:
            type: object
            properties:
              input_tokens:
                type: integer
              prompt_tokens:
                type: integer
              completion_tokens:
                type: integer
              total_tokens:
                type: integer
    QueryResponse:
      type: object
      properties:
        queryId:
          type: integer
        answer:
          type: string
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        graphInfo:
          type: object
          nullable: true
          properties:
            entities:
              type: array
              items:
                type: string
            relations:
              type: array
              items:
                type: object
                properties:
                  source:
                    type: string
                  target:
                    type: string
                  type:
                    type: string
        contradictions:
          type: array
          items:
            $ref: '#/components/schemas/Contradiction'
        tokenStats:
          $ref: '#/components/schemas/TokenStats'
        responseTime:
          type: number
        confidence:
          type: string
          enum: [high, medium, low]
        model:
          type: string
        timestamp:
          type: string
          format: date-time
    QueryHistoryItem:
      type: object
      properties:
        id:
          type: integer
        novelId:
          type: integer
        query:
          type: string
        answer:
          type: string
          description: 简短摘要（前100字）
        model:
          type: string
        totalTokens:
          type: integer
        confidence:
          type: string
        createdAt:
          type: string
          format: date-time
        feedback:
          type: string
          enum: [positive, negative]
          nullable: true

    # ========== 图谱相关 ==========
...
    TokenStatsReport:
      type: object
      properties:
        totalTokens:
          type: integer
        queryCount:
          type: integer
        indexCount:
          type: integer
        averageTokensPerQuery:
          type: integer
        byModel:
          type: object
          additionalProperties:
            type: object
            properties:
              totalTokens:
                type: integer
              promptTokens:
                type: integer
              completionTokens:
                type: integer
              apiCalls:
                type: integer
        byOperation:
          type: object
          properties:
            index:
              type: object
              properties:
                totalTokens:
                  type: integer
                count:
                  type: integer
            query:
              type: object
              properties:
                totalTokens:
                  type: integer
                count:
                  type: integer
        dailyStats:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              totalTokens:
                type: integer
              queryCount:
                type: integer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT认证（可选，个人应用可不启用）

# 默认不启用安全认证（个人应用）
# security:
#   - BearerAuth: []